+++ crates/lz78/src/util.rs
@@ -3,10 +3,10 @@ use ndarray::Array1;
 
 /// Given a PDF and a sample of a Uniform([0, 1)) random variable, randomly
 /// sample a number from 0 to pdf.len() - 1 as per the PDF.
-pub fn sample_from_pdf(pdf: &Array1<f64>, sample: f64) -> u64 {
+pub fn sample_from_pdf(pdf: &Array1<f32>, sample: f32) -> u64 {
     let cdf = pdf
         .iter()
-        .scan(0.0_f64, |sum, i| {
+        .scan(0.0_f32, |sum, i| {
             *sum += i;
             Some(*sum)
         })
